stop overloading nil with scalar serialization
It's perfectly valid in some scenarios for a scalar to deserialize to
null, for example when passing a Void type around. Previously this would
be treated like an error. Now we only treat actual errors like errors.
In practice not having an error return value led to a lot of panicky
Scalar implementations, so this feels more correct anyway. It also lets
you return a helpful error message instead of leaving you guessing.