check whether object implements interface at runtime
The pre-existing implementation of the graphql server required that
objects essentially pre-register the interfaces they implement when they
are added to the server and would error out if the object is attempted
to be used when it wasn't registered as an interface.

This makes sense for general graphql use cases, but in ours we don't
want users to have to say ahead of time which interfaces they are
implementing; they may not even know of the interfaces they could be
implementing.

While it's possible that we could add support for that in the future, as
a baseline we don't want to require it.

So this change re-uses the existing logic for checking whether an object
implements an interface (normally only called when the schema is
initialized) but calls it now at the time the object is returned as an
interface, erroring out if it doesn't match.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>